Rework survey
